+++ src/GSvar/DiagnosticStatusOverviewDialog.cpp
@@ -92,7 +92,7 @@ void DiagnosticStatusOverviewDialog::updateOverviewTable()
 			tmp.append(entry.disease_info);
 		}
 		addItem(r, 5, tmp.join(", "));
-		addItem(r, 6, sample_data.phenotypesAsStrings().join(", "));
+		addItem(r, 6, sample_data.phenotypes.toString());
 		addItem(r, 7, processed_sample_data.run_name);
 		addItem(r, 8, processed_sample_data.processing_system);
 		addItem(r, 9, processed_sample_data.project_name);