+++ src/GSvar/VariantDetailsDockWidget.cpp
@@ -30,10 +30,17 @@ VariantDetailsDockWidget::VariantDetailsDockWidget(QWidget* parent)
 	connect(ui->gnomad, SIGNAL(linkActivated(QString)), this, SLOT(gnomadClicked(QString)));
 	connect(ui->var_btn, SIGNAL(clicked(bool)), this, SLOT(variantButtonClicked()));
 	connect(ui->trans, SIGNAL(linkActivated(QString)), this, SLOT(transcriptClicked(QString)));
+	connect(ui->som_details_prev, SIGNAL(clicked(bool)), this, SLOT(previousSomDetails()));
+	connect(ui->som_details_next, SIGNAL(clicked(bool)), this, SLOT(nextSomDetails()) );
 
 	//set up transcript buttons
 	ui->trans_prev->setStyleSheet("QPushButton {border: none; margin: 0px;padding: 0px;}");
 	ui->trans_next->setStyleSheet("QPushButton {border: none; margin: 0px;padding: 0px;}");
+
+	//set up somatic detail buttons
+	ui->som_details_prev->setStyleSheet("QPushButton {border: none; margin: 0px;padding: 0px;}");
+	ui->som_details_next->setStyleSheet("QPushButton {border: none; margin: 0px;padding: 0px;}");
+	enableSomDetailsArrows();
 
 	//set up content label properties
 	QList<QLabel*> labels = findChildren<QLabel*>();
@@ -95,6 +102,16 @@ void VariantDetailsDockWidget::setLabelTooltips(const VariantList& vl)
 	ui->label_ngsd_group->setToolTip(vl.annotationDescriptionByName("NGSD_group", false).description());
 	ui->label_ngsd_comment->setToolTip(vl.annotationDescriptionByName("comment", false).description());
 	ui->label_ngsd_validation->setToolTip(vl.annotationDescriptionByName("validation", false).description());
+
+	//somatic details
+	ui->label_somcgi_driver_status->setToolTip( vl.annotationDescriptionByName("CGI_driver_statement", false).description() );
+	ui->label_cgi_gene_role->setToolTip( vl.annotationDescriptionByName("CGI_gene_role", false).description() );
+	ui->label_som_class->setToolTip( vl.annotationDescriptionByName("somatic_classification", false).description() );
+	ui->label_somatic_count->setToolTip( vl.annotationDescriptionByName("NGSD_som_c", false).description() );
+	ui->label_somatic_oncogene->setToolTip(vl.annotationDescriptionByName("ncg_oncogene", false).description() );
+	ui->label_somatic_tsg->setToolTip( vl.annotationDescriptionByName("ncg_tsg", false).description() );
+	ui->label_somatic_cancerhotspots->setToolTip( vl.annotationDescriptionByName("CANCERHOTSPOTS_ALT_COUNT", false).description() );
+	ui->label_somatic_cmc_class->setToolTip( vl.annotationDescriptionByName("CMC_mutation_significance", false).description() );
 }
 
 void VariantDetailsDockWidget::updateVariant(const VariantList& vl, int index)
@@ -203,6 +220,9 @@ void VariantDetailsDockWidget::updateVariant(const VariantList& vl, int index)
 	setAnnotation(ui->somatic_oncogene, vl, index, "ncg_oncogene");
 	setAnnotation(ui->somatic_tsg, vl, index, "ncg_tsg");
 
+	setAnnotation(ui->somatic_cancerhotspots, vl, index, "CANCERHOTSPOTS_AA_CHANGE");
+	setAnnotation(ui->somatic_cmc_class, vl, index, "CMC_mutation_significance");
+
 	//update NGSD button
 	ui->var_btn->setEnabled(LoginManager::active());
 }
@@ -290,7 +310,7 @@ void VariantDetailsDockWidget::setAnnotation(QLabel* label, const VariantList& v
 				else if (entry.details.contains("CLASS=DM")) color = RED;
 				else if (entry.details.contains("CLASS=DP") || entry.details.contains("CLASS=DFP")) color = ORANGE;
 
-				text += formatLink(entry.id, "https://portal.biobase-international.com/hgmd/pro/mut.php?acc=" + entry.id, color) + " ";
+				text += formatLink(entry.id, "https://my.qiagendigitalinsights.com/bbp/view/hgmd/pro/mut.php?acc=" + entry.id, color) + " ";
 				tooltip += nobr() + entry.id + ": " + entry.details;
 			}
 		}
@@ -504,6 +524,20 @@ void VariantDetailsDockWidget::setAnnotation(QLabel* label, const VariantList& v
 		else if(name=="ncg_tsg")
 		{
 			if(anno.contains("1")) text = formatText(anno, RED);
+			else text = anno;
+		}
+		else if(name=="CANCERHOTSPOTS_AA_CHANGE")
+		{
+			int i_total_mut_cout = vl.annotationIndexByName("CANCERHOTSPOTS_TOTAL_MUT", true, false);
+			int i_alt_count = vl.annotationIndexByName("CANCERHOTSPOTS_ALT_COUNT", true, false);
+			if(i_total_mut_cout >= 0 && i_alt_count >=0 && !anno.isEmpty())
+			{
+				text = anno.append(":").split(":")[1] + " (" + vl[index].annotations()[i_alt_count] + "/" + vl[index].annotations()[i_total_mut_cout] + ")";
+			}
+		}
+		else if(name=="CMC_mutation_significance")
+		{
+			if(anno == "1" || anno == "2" || anno == "3") text = formatText(anno, RED);
 			else text = anno;
 		}
 		else if(name=="MMSplice_DeltaLogitPSI")
@@ -824,6 +858,24 @@ void VariantDetailsDockWidget::variantButtonClicked()
 	if (variant_str.isEmpty()) return;
 
 	emit openVariantTab(Variant::fromString(variant_str));
+}
+
+void VariantDetailsDockWidget::nextSomDetails()
+{
+	ui->somaticDetailsWidget->setCurrentIndex(ui->somaticDetailsWidget->currentIndex()+1);
+	enableSomDetailsArrows();
+}
+
+void VariantDetailsDockWidget::previousSomDetails()
+{
+	ui->somaticDetailsWidget->setCurrentIndex(ui->somaticDetailsWidget->currentIndex()-1);
+	enableSomDetailsArrows();
+}
+
+void VariantDetailsDockWidget::enableSomDetailsArrows()
+{
+	ui->som_details_prev->setEnabled(ui->somaticDetailsWidget->widget(ui->somaticDetailsWidget->currentIndex()-1) != nullptr);
+	ui->som_details_next->setEnabled(ui->somaticDetailsWidget->widget(ui->somaticDetailsWidget->currentIndex()+1) != nullptr);
 }
 
 QList<KeyValuePair> VariantDetailsDockWidget::DBEntry::splitByName() const